Back conectado y funcionando con base de datos

+++ src/main/java/com/example/coworkingprojectback/repository/ImagenRepository.java
@@ -8,5 +8,5 @@ import java.sql.Blob;
 import java.util.Optional;
 
 @Repository
-public interface ImagenRepository extends JpaRepository <Imagen, Long> {
+public interface ImagenRepository extends JpaRepository <Imagen, Integer> {
 }
